migration: Add prometheus metric to track latency
Add a prometheus metric so we can track the latency for each migration.
We also add a name field to the migration, to ensure we can add the
migration name to the metric labels.

Also add a log for successful migration, since adding relative path to
the metric would blow up cardinality, we add a log to retain repository
level information.