migration: Pass migrations down to the manager
The migration manager implicitly sets the migrations by using the global
migrations variable. Let's make this explicit, this way, we can
initialize the migrations with the necessary information and then pass
those migrations to the manager, this way, the manager doesn't have to
deal with dependencies of the manager.

Also make `NewPartition()` a private function, since it doesn't need to
be exposed outside of the package, external users should directly use
the factory provided.