Move optimization related tests to housekeeping/manager
We already have some optimization related tests in housekeeping/manager,
this commit moves the rest of the tests and remove the duplicate ones.
Since internal details of the repository is not exposed through API,
these tests don't run when WAL is enabled. A different approach would be
to test against the collected data structures metrics, which would be
handled in a separate MR.