Use ForceSnapshotInvalidation and ForceWALSync
To ensure the repository is fully written to disk before performing any
operations, by default we use the ForceSnapshotInvalidation workaround
in gittest.CreateRepository. This guarantees that subsequent read
requests access a fresh snapshot, if a read request were used for WAL
sync, the snapshot would be cached, and follow-up read requests would
use an outdated snapshot. This is particularly problematic when raw Git
operations, such as the ones present in gittest.WriteCommit, modify the
repository state in the meantime.