gitlab: Pass RPC context to Allowed API call
When checking whether a change is allowed to go through we use
/internal/allowed API endpoint from GitLab Rails. One of the
parameters for this call is "action" which is currently hardcoded
to "git-receive-pack". We want to enrich this with actual RPC
that triggered the allowed call, but instead of hardcoding the
RPC names, we decided to pass metadata from clients that Gitaly
will directly relay to internal API. This allows for more flexible
and detailed information exchange, and can be easily extended
without needing to upgrade Gitaly each time.