Remove snapshotting from pack-refs verification
`pack-refs` operations may conflict with reference updates.
Reference deletions lead to an abort, but creations and updates
are instead resolved by dropping the conflicting loose ref removals
that `pack-refs` would perform. Currently this verification is done
against the staging snapshot. Snapshotting is slow, so we'd like to
perform this without a snapshot.

This commit performs the verification directly against the target
repository. Overall the logic remains the same but we can't delete
files or directories from the target repository during verification
which happens before the transaction commits. To workaround this, we
instead keep track of the deleted files and directories in memory.
This allows us stop using a snapshot for this operation.

Reftables are only reading the current 'tables.list' file. We read
that now directly from the target repository instead of reading it
from the staging snapshot. As the staging snapshot is no longer
needed, we remove it.