Use staging snapshot for pruning verification
When we're committing a transaction that prunes objects, we need to
verify none of the pruned objects were concurrently made reachable.
This is done by applying the result of the pruning run into a snapshot
and checking whether the objects still exist. We're currently creating
a snapshot and cleaning it up ad-hoc rather than using the staging
repository. This duplicates code but also leads to cleaning up the
snapshot during the critical section rather than deferring the clean
up to a background goroutine as done with the staging snapshot.

Replace the custom snapshotting code with the staging snapshot
functionality.