Release mutex as soon as possible
verifyHousekeeping is deferring the mutex unlock. The mutex is meant
to cover multiple accesses in to walkCommittedEntries done in the
housekeeping functions as client goroutines may commit/rollback and
concurrently modify the state in the committedEntries list.

Move the mutex handling inside the walkCommittedEntries function.
This way it spans the data that is accessed but the mutex is released
as soon as possible.