hooks: Pass client context to hook payload
Rails needs to send some information to internal endpoints via Gitaly,
and the request first goes through git and back to Gitaly via Hooks
before Gitaly finally calls the internal endpoints such as "allowed".
Therefore, we need to first add the context information to the hook
payload to be able to access it later.