migration: Cancel context when partition is stopped
The migration manager wraps around a storage partition. It instantiates
running the migrations when the Begin function of the partition is
called. However, if the Close function of the partition is called while
migrations are run, we don't propagate this information to the loop
which runs the migrations.

Let's add an isolated context and cancellation function to the manager,
so that we can cancel the context as needed.