git: Set "followremotehead" to "never" for fetches
https://lore.kernel.org/git/20240910203835.2288291-1-bence@ferdinandy.com/
was recently merged into the `next` branch resulting in some failures in
our nightly tests.

Set the "followremotehead" option to revert to the previous behaviour so
that the tests pass and we're not surprised by any changes during
fetches. We'll also investigate whether we can just adopt the new
behaviour rather than manually setting the local HEAD after a fetch.