Demonstrate pack-objects cache misses with transactions
Pack-objects cache currently doesn't work well with transactions.
We're not correctly using the original relative path of the repository
in the cache key. This leads to transactions with different snapshots
to compute a different key which leads to cache misses.

Our current tests did not catch this. Expand the packObjectsHook tests
in smarthttp and ssh services to cover this. The tests were for
some reason stubbing out the entire hook flow. Change the test cases
to actually cover the entire hook flow. We wrap the underlying
pack-object cache in the server and record the cache keys accessed.
This asserts that the pack-objects cache was actually called, and
allows us to assert that the keys correctly match regardless of the
transaction snapshot.