storage: Remove StateDirectory() access from log.Manager
This method is used by a TransactionManager's test only. This function
exposes the internal state of the log manager. Although `log` is a
sub-package of the `partition``,` this exposure is a leaky abstraction.
This commit removes the access to the `StateDirectory()` method. The
caller test now asserts the existence and content of LSN directories
instead.