storage: Create storage.LogManager interface to abstract log.Manager
The `storage.LogManager` interface was introduced to abstract out the
interaction between log consumers and `log.Manager`. It contains a small
portion of public methods of `log.Manager`.

This commit renames that interface to LogReader and repurpose LogManager
interface for the full management ability. TransactionManager now
interacts with the log manager via this interface. This step is a
pre-requisite for Raft integration.

Some next commits will create a Raft manager that implements the
`storage.LogManager` interface. This approach allows us to switch
between log managers with/without Raft enablement.