Restore transaction into context in PackObjectsHook handler
pack-objects cache has currently a bad hit rate with transactions.
This is because we're not correctly using repository's original
relative path in the cache key but we're including the rewritten
relative path that contains the snapshot prefix. This means that
all fetch calls targeting the same repository from different
snapshots will miss the cache entries created from other snapshots.

While we already have logic to extract the original relative path
in place, we're not correctly restoring the transaction into the
context when receiving a call to the pack objects hook handler.
Typically RPC handlers would already have the transaction in the
context but because the hook calls cross process boundaries and come
back as different RPC invocations, the context does not carry over.