Remove GITALY_USE_SIDECHANNEL flag in gitaly-ssh
SSHUploaddPack was deprecated for a fairly long time. It was removed
from all downstream dependencies of Gitaly. It's time to remove this
RPC completely from Gitaly.

In a prior commit, GITALY_USE_SIDECHANNEL is always set to true from all
code paths. This commit removes the check of that environment variable
in gitaly-ssh. The callers don't have other options than using
sidechannel now.