Remove SSHUploadPack RPC from SSH service handler
Since all internal clients and external clients were removed, it's safe
to remove this RPC from the service handler. After the removal, all
relevant tests are directed to SSHUploadPackWithSidechannel instead.

Other RPCs in the same RPC service use a non-sidechannel dialer. In
theory, those RPCs will work fine on a sidechannel connection. However,
it is verified in the sidechannel package. No need to assert them in the
service package.