gitaly-server: Add localrepo.Factory as a new server dependency
In order to use the localrepo.Factory throughout Gitaly's
services, we first need to add an instance of the Factory
to the list of dependencies of Gitaly services. That way
the Factory can be used by the multiple services that
requires it.

This commit only adds the Factory to the list of dependencies;
it does not add any logic which uses this dependency yet.
Those changes will come in a subsequent commit.

This commit also updates the various *_test.go files
that create a test Gitaly server, in order to create
a test server which includes the Factory as well.