services: Use localrepo.Factory throught the Gitaly services
The changes introduced uses the localrepo.Factory of each Gitaly's
services, which were added in a previous commit.

This refactoring aims to enhance the readability of the code by making
use of an already existing Factory in order to reduce the number of
dependencies that needs to be passed along the various layers of the
code in order to create a new instance of localrepo.Repo.

Those changes were first proposed and discussed in the following issue:

* https://gitlab.com/gitlab-org/gitaly/-/issues/6383