storage: Inject the log manager into TransactionManager
Priorly, `TransactionManager` creates an instance of `log.Manager` in
the initializer method. In the initializer, the log consumer is passed
as a parameter, although the transaction manager doesn't need it. In
some prior commits, the `log.Manager` was replaced by the abstract
`storage.LogManager` interface.

This commit modifies the initializer method of the `TransactionManager`
to receive an instance of `storage.LogManager`. This change serves two
purposes:
* This detaches the connection between the log manager and transaction
  manager completely. The log consumer now interacts with the log
  manager exclusively.
* Later, we'll swap in a log manager with Raft functionalities.

This approach simplifies the interaction between TransactionManager and
RaftManager: Transaction Manager -> Distributed Log Manager (Raft) ->
Local Log Manager. The distributed log manager and local log manager
share the same interface (`storage.LogManager`). As soon as the log
manager confirms a log entry is appended, the transaction manager is
allowed to apply that log entry. All networking activities, especially
committing entries to replicas, are out of the concerns of
TransactionManager.