transaction: Fix race condition
There are some flaky tests related to transactions which are failing
when we try to simulate a crash. They sometimes fail at the commit stage
expecting the ErrTransactionProcessingStopped error, and ideally we
would want the failure to happen later. To fix this we can remove the
<-mgr.closed case as transaction.result channel should return the
appropriate error. For more info, refer this discussion -
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/7515#note_2264741276