offloading: Add sink for offloading storage
The offloading storage is a cloud object store where we can save
our offloaded object to. As part of repository offloading,
interacting with the offloading storage is required for operations
such as downloading, uploading, deleting, and listing.

This commit introduces an new sink implementation to serve as an
abstraction where we can interact with the offloading storage.