offloading: Add a mock bucket for simulation delay response
We have retry and cancel support in the offloading sink. In order
test those, we have to mock a bucket that can simulate delay and error.

In this commit, we add a mock bucket. The bucket is based on fileblob
bucket but intercept its response with cuostizable delay and error.
With such simulation, we can write test cases to examine the
offloading sink's behavior under long delay and frequent error.