storage: Allow log manager returns an error via the notifcation queue
The log manager exposes a notififcation queue to communicate with the
transaction manager. The queue uses an empty struct to notify any
changes now. In some next commits, we'll need to let the transaction
manager know when the log manager failed. This commit updates the
signature of notifcation queue to an error channel.