storage: Unify TransationManager creation to via partition.Factory
Recently, there are two ways to create an instance of the
`TransactionManager`: use `partition.Factory` or create it directly. The
former way is used in server setup, while the second one is used in the
test only. This derivation adds a subtle but dangerous difference
between the test setup and real production code, especially when we
would like to add some more logic to the factory in upcoming commits. It
also adds some verbosity to the test setup.

This commit unifies the TransactionManager setup. `partition.Factory` is
used universally in all occurrences in the test and the server setup.