Remove bundle generation logic from the Sink
This commit is the first of a serie of refactor in the
bundle-uri package. The end goal of the refactor is to
put all the bundle-generation logic into the bundle-uri
manager, and stop calling the Sink directly. This will
streamline the code both for the bundle fetch and
the bundle generation operations.