SignedURL method moved from Sink to Manager
By moving this method, the Manager now holds all
logic to manage a bundle. This removes the need,
for any Service, to hold a reference to the Sink
used internally by the Manager.

Before this change, managing a Bundle was a bit
confusing. We had to use the Manager to generate
the bundle, but we also had to use the underlying
Sink, which is part of the intertanl implementation
of the Manager, to get the SignedURL of the bundle.

Now, a service can perform all operations
needed to manage a bundle solely by using
the Manager.