raft: Implement gRPC transport interface
This commit introduces a gRPC transport interface for handling Raft
communication. It implements the logic for sending and receiving Raft
messages, including packing and unpacking log data. Additionally, the
RaftManagerRegistry facilitates Raft manager registration and partition
lookup.
For the address lookup, we introduce a static routing table for now. In
future, the table can be dynamic so that the addresses are propagated
through gossiping.