gitaly: Optimize LSN range output
Previously we have implemented the status command to print out all
available WAL entries, but this can produce massive amount of lines
in a production environment. With this change we aim to improve the
readability of the output and make it more useful by printing only
the ranges of contiguous LSN sequences.