migration: Add reftable migration to dry-run
Add a new migration to migrate a repository from files backend to the
reftable backend. This is done using the `git-refs-migrate(1)` command,
which converts a given repository to a specified backend.

To ensure integrity of the repository, we compare the hash for all the
refs before and after the migration.

Finally, we add the migration to the dry-run set of migraitons,
eventually it will be moved to the actual set of migraitions.