storage: Create snapshot locks remote lazily
TransactionManager maintains a hashmap of snapshot locks used for
synchronizing snapshotters with the log application. This system is used
in two places:
- Before a transaction starts, it is blocked until the log entry of its
  snapshot is applied.
- Before a log entry is applied, it waits until all transactions who
  depend on the prior log entry complete initialization.

Recently, the snapshot lock is created when a log entry is appended.
However, this setup step is handled by a private method of
TransactionManager. If a log entry is inserted via a different way, the
respective snapshot lock is not created. Thus, following transactions
panic.

This commit lets TransactionManager create snapshot locks lazily before
they are used.