backup: Convert the cleanup functions list to a stack
In the recovery sub-command, we set up a recovery context that contains a
list of cleanup functions. Functions in this list are executed in FIFO
order. This execution turns out to be wrong because the latter
declaration depends on former dependencies. While cleaning up resources
in the storage manager, the DB is already closed. It leads to occasional
"DB closed" error.

This commit converts the cleanup function list to a stack so that
cleanup functions are executed in LIFO order.