backup: Confirm recovered log entries are being applied
We append all the archived WAL entries into a partition, but there is no
guarantee they are applied before exiting. They will be applied next
time the transaction manager runs (after a restart, for example). If a
log entry cannot be applied for any reason (broken, wrong bucket, etc.),
the user will find out, but it requires an in-depth investigation. Until
the reason is exposed, that partition is always in a broken state.

As a result, the recovery tool should better report failed log
applications to the user, sound and clear. The broken log entry should
also be deleted from WAL. It's up to the user's decision to make the
next course of action.