storage: Remove fsync system call after writing the manifest
The TransactionManager previously triggered an fsync system call after
writing the manifest file to ensure data was flushed to disk. This was
necessary because some files were staged only within the critical
section due to the lack of robust conflict checks. This limitation
prevented staging the entire log entry before queuing it for commit.

Following the resolution of
https://gitlab.com/gitlab-org/gitaly/-/issues/5892, conflict checking
was significantly improved. We implemented a comprehensive file-level
conflict check, combined with logical checks, enabling all files of a
WAL log entry to be staged before admission. Processing a transaction no
longer requires reading staged files, as we now rely on in-memory
filesystem records for this purpose.

At a later stage, when the log.Manager moves the log entry from the
staging directory to the committed directory, another fsync is
triggered. If the process crashes before this point, the transaction
remains uncommitted, ensuring data integrity.

Therefore, the fsync call after writing the manifest is no longer
necessary. Removing it does not compromise data consistency.