housekeeping: Generalize Git config test
In Git 2.48.0, the order of Git config entries has changes for newly
initialized repositories. In the housekeeping package the
`TestRepositoryManager_CleanStaleData_unsetConfigurationTransactional`
test relies on the exact order for the Git config entries it validates.

Update the test to no longer care about the order of entries in the Git
config.