repository: Adjust config generation for transaction test
With the release of Git v2.48, the order of Git config entries written
by git-init(1) and git-clone(1) has changed slightly. While the
difference is only in the order of config entries, transaction votes
rely on the content of files in the repository. Any difference results
in a different votes.

The `TestCreateRepositoryFromBundle_transactional` test validates that
transaction vote generation for the `CreateRepositoryFromBundle` RPC
works are expected. The expected vote is determined from a repository
used to create the bundle which was created with git-init(1). The actual
vote is determined from a repository created from this bundle via
git-clone(1). Due to the recent Git change, the configs no longer
exactly match which will result in test failure when using Git v2.48.

Adapt the test to avoid these config differences by just using the
actual config in the computed transction vote hash.