ci: adapt one of the tests to use Meson
Adapt one of the tests to use Meson to build Git instead of using the
Makefile. For now we'll retain jobs that build with both systems until
we can completely phase out the Makefile-based jobs. This requires us to
bump the minimum required Git version first though so that all supported
Git versions have Meson available.