go-toolchain: Downgrade gitaly-init-cgroups version
The gitaly-init-cgroups tool has an inconsistent
'toolchain' version in the go.mod file. All other
tools are on 1.22.6, but this one was on 1.22.8.

This MR downgrades the 'toolchain' version of this
tool to 1.22.6, so it is consistent with all other
tools and binaries of the project.

This change is also motivated by the fact that having
different go.mod files with inconsistent `toolchain'
versions across the same project can cause some version
conflicts when people uses IDE, which prevents building
and debugging.