localrepo: Disable `CloneBundle()` fsck
Starting in Git version 2.46.0, executing git-clone(1) on a bundle
performs fsck checks when `transfer.fsckObjects` is enabled. Prior to
this, this configuration was always ignored and fsck checks were not
run. Unfortunately, fsck message severity configuration is ignored by
Git only for bundle clones.

This issue is resolved in Git version 2.48.0, but until this gets rolled
out to GitLab, disable `transfer.fsckObjects` so bundles containing fsck
errors can continue to be cloned. This matches behavior prior to Git
version 2.46.0.