fix(range-diff): fix bad range diff line regex pattern
The regular expression for parsing diff lines in `range-diff`
does not correctly account for multiple spaces at the beginning
or within the lines. This can lead to parsing errors when the
`range-diff` output exceeds 10 diff lines. The issue can be
resolved by modifying the regular expression to properly support
parsing these spaces.

Signed-off-by: ZheNing Hu <adlternative@gmail.com>