recovery: Fix WAL entry tar extraction path
We download the archived WAL entry, extract it to a temporary dir and
then finally move it to the partition's log directory. Currently this
results in repeated LSN number in the final path such as:
"/.../+gitaly/partitions/d4/73/2/wal/0000000000005/0000000000005/".

This happens because we are extracting the tar into a path such as
"/tempDirectory/0000000000005" but the tar file itself also contains
the directory information such as:

➜  tar -tvf 0000000000005.tar
drwxrwsrwx  0 git    git         0 Jan 22 12:28 0000000000005/
-rw-rw-rw-  0 git    git        41 Jan 22 12:28 0000000000005/1
-rw-rw-rw-  0 git    git       830 Jan 22 12:28 0000000000005/MANIFEST

And then finally we move it to its final destination in log with path
like "../wal/0000000000005" using os.Rename(srcPath, destPath), which
then results in final path with repeated LSN number.

Changes in this MR:
- Extracted the tar into a staging directory and then validated that
the log entry exists at the correct path. Afterwards passed the path
of the log entry directly to the log manager which fixes the issue.
- Update unit tests to use realistic tar file structure.