Bundle-URI: Strategy for auto-generation of bundle
When generating a bundle, the manager increment a progress
tracker and compares the number of `in progress` requests
with a given threshold. If the threshold is reached, a
bundle is generated. That bundle is generated into it's
own goroutine. Once the comparison is done, the progress
tracker is decremented. Those successive operations are
happening one after another:

1. Increment counter
2. Compare
3. Decrement

The fast pace at which those operations are executed makes
it very unlikely that a bundle will be generated. The real
issue is that the decrement of the counter happens right after
the comparison. In such an implementation, a repository would
need to be cloned `threshold` number of times in a matter of
milliseconds or less in order for the counter to reach the
threshold.

This commit introduces a new `Strategy` interface that could be
used by the manager to delegate the decision whether a bundle
should be generated or not. It also introduces an implementation
of that `Strategy` interface that tracks the number of occurrences
for which a clone was requested for a repository. If the number
of occurrences reaches a given threshold within an given interval
of time, then a bundle is generated. It also keeps track of the
time a bundle was generated for a repository, to avoid creating
a bundle that was generated not so long ago. Finally, it limits
the number of concurrent bundle generation that can occur at
the same time.