Integrating the Strategy interface within the manager
A previous commit added a Strategy interface to handle
the logic that determines whether a bundle should be
generated or not. Although that previous commit did
introduce that interface, it wasn't used.

This commit updates the bundleURI manager to use this
interface, using the OccurrencesStrategy implementation.

It also updates the various tests that needed to be
changed.