mod: Upgrade badger version
We haven't updated badger dependency for a while and the version
we are currently using has vulnerability issue. We are upgrading
the badger version to latest to fix this.
This new version of badger iterator panic if Next is called on
exhausted iterator, thus we also introduced a validity check for
the iterator before calling Next.