ci: override HOME when running tests unprivileged
We use setpriv(1) to run tests as an unprivileged user, which is done
because "root" has special permissions that might override for example
filesystem permissions. Consequently, some tests that rely on these
would otherwise fail.

We don't override the environment with setpriv(1) though. This is mostly
working as designed because we indeed want to propagate most of the
environment variables to the underlying tests. But for other environment
variables it may cause problems. One such variable is "HOME", which
continues to point to the home directory of the "root" user. So any
executables that may try to read something from the home directory may
fail.

This is the case with Meson, which tries to read several files from the
user's home. Naturally, it gets an error due to a lack of permissions
when trying to do so, which makes it fail.

Fix this issue by overriding HOME to point to a temporary directory.